refactor(nvim): Modularized nvim plugin config
This commit modularizes the old nvim configuration using two ways.
If the environment is managed by Home Manager, then a special value is set by Home Manager:

                           vim.g.home_manager = true

With this value being present and set to true, the module `plugin-loader.lua` iterates
through all files in the `plugins/` directory and attempts to call the configure
function.

If this value is missing or set to false, however, then the package manager `lazy.nvim`
is being bootstrapped which in turn installs all plugins and calls the configure function
by itself.

Signed-off-by: Julian Kirsch <contact@juliankirsch.me>